extracted login method from Logic to LoginController
