+++ src/main/java/backend/Logic.java
@@ -1,6 +1,17 @@
 package backend;
 
-import static util.Futures.withResult;
+import backend.resource.Model;
+import backend.resource.MultiModel;
+import github.TurboIssueEvent;
+import org.apache.logging.log4j.Logger;
+import org.eclipse.egit.github.core.Comment;
+import prefs.Preferences;
+import ui.UI;
+import util.Futures;
+import util.HTLog;
+import util.Utility;
+import util.events.RepoOpenedEvent;
+import util.events.testevents.ClearLogicModelEventHandler;
 
 import java.time.LocalDateTime;
 import java.time.ZoneId;
@@ -8,19 +19,7 @@ import java.util.*;
 import java.util.concurrent.CompletableFuture;
 import java.util.stream.Collectors;
 
-import github.TurboIssueEvent;
-import org.apache.logging.log4j.Logger;
-
-import backend.resource.Model;
-import backend.resource.MultiModel;
-import org.eclipse.egit.github.core.Comment;
-import prefs.Preferences;
-import ui.UI;
-import util.Futures;
-import util.HTLog;
-import util.Utility;
-import util.events.testevents.ClearLogicModelEventHandler;
-import util.events.RepoOpenedEvent;
+import static util.Futures.withResult;
 
 public class Logic {
 
@@ -31,9 +30,7 @@ public class Logic {
     private final Preferences prefs;
 
     private RepoIO repoIO;
-
-    // Assumed to be always present when app starts
-    public UserCredentials credentials = null;
+    public LoginController loginController;
 
     public Logic(UIManager uiManager, Preferences prefs, boolean isTestMode, boolean enableTestJSON) {
         this.uiManager = uiManager;
@@ -41,6 +38,7 @@ public class Logic {
         this.models = new MultiModel(prefs);
 
         repoIO = new RepoIO(isTestMode, enableTestJSON);
+        loginController = new LoginController(repoIO);
 
         // Only relevant to testing, need a different event type to avoid race condition
         UI.events.registerEvent((ClearLogicModelEventHandler) e -> {
@@ -68,15 +66,6 @@ public class Logic {
 
     private CompletableFuture<Boolean> isRepositoryValid(String repoId) {
         return repoIO.isRepositoryValid(repoId);
-    }
-
-    public CompletableFuture<Boolean> login(String username, String password) {
-        String message = "Logging in as " + username;
-        logger.info(message);
-        UI.status.displayMessage(message);
-
-        credentials = new UserCredentials(username, password);
-        return repoIO.login(credentials);
     }
 
     public void refresh() {
